requirements.txt
django
djangorestframework
psycopg2-binary
python-dotenv
celery
redis 
transformers
sentencepiece
torch==2.3.1
django-redis
gunicorn